+++ R/legend.R
@@ -580,6 +580,7 @@ addLegendNumeric <- function(map,
                              fillOpacity = 1,
                              ...) {
   stopifnot( attr(pal, 'colorType') == 'numeric' )
+  shape <- match.arg(shape)
   rng <- range(values, na.rm = TRUE)
   breaks <- pretty(values, bins)
   if ( breaks[1] < rng[1] ) {
@@ -919,6 +920,10 @@ addLegendFactor <- function(map,
 #'
 #' the values used to generate color from the palette function
 #'
+#' @param numberFormat
+#'
+#' formatting functions for numbers that are displayed e.g. format, prettyNum
+#'
 #' @param ...
 #'
 #' arguments to pass to
@@ -1007,6 +1012,7 @@ addLegendSize <- function(map,
                           fillOpacity = opacity,
                           breaks = 5,
                           baseSize = 10,
+                          numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
                           ...) {
   shape <- match.arg(shape)
   sizes <- sizeBreaks(values, breaks, baseSize)
@@ -1035,7 +1041,7 @@ addLegendSize <- function(map,
                  opacity = opacity,
                  fillOpacity = fillOpacity,
                  `stroke-width` = strokeWidth)
-  addLegendImage(map, images = symbols, labels = names(sizes),
+  addLegendImage(map, images = symbols, labels = numberFormat(as.numeric(names(sizes))),
                  title = title, labelStyle = labelStyle,
                  orientation = orientation, width = sizes, height = sizes, ...)